Update __init__.py
During initial setup and baseline, emails for every alert are overwhelming.  Perhaps starting with the "New Sweet Security Alert" email.emailUser line commented out would allow for baselining, then remove comment once normalized/tuned.

Would be fine as-is for an on-prem smtp server, but not so much for gmail, etc.